+++ release_notes.md
@@ -25,6 +25,7 @@ PREESM Changelog
 * SPiDER codegen considers empty PAPIFY configs as papify=false;
 * Update round buffer memory script to handle multiple inputs;
 * Schedule: Add an interface to perform transform
+* PiSDF: Add more rules to check mergeability, remove getPreds/Succs from TopologyHelper
 * Codegen:
   * Add IteratedBuffer, ClusterBlock, SectionBlock for clustering codegen
   * Add conditional "#pragma omp parallel for" on FiniteLoopBlock